Filter WARN and ERROR levels to stderr
This updates the log4j and logback configurations to send
WARN and ERROR to stderr, while still sending INFO to stdout.

It also tweaks the log4j logging configuration to more precisely match
the logback configuration in the INFO case, so that an "[INFO] " prefix
is not shown for either logging mechanism.

This change will be very useful for differentiating between different
log levels in a simpler way; in particular, ImageJ can send both stdout
and stderr to a Console window, coloring stderr messages in red and
setting the Console to visible when they occur, but not displaying the
Console by default when only stdout messages occur (because many ImageJ
plugins spam stdout with lots of debugging output).

Thanks to Uriah Carpenter and Jonas Berlin for their SO answers:
* http://stackoverflow.com/a/5653532
* http://stackoverflow.com/a/10314912